refactor: enhance GPU info logging and remove types file
